Change folder structure: don't have a java folder for 1 class
