+++ OODebug/src/compiler/JavaCompiler.h
@@ -26,39 +26,35 @@
 
 #pragma once
 
-#include "../../oodebug_api.h"
+#include "../oodebug_api.h"
 
 namespace Model {
 	class TreeManager;
 	class Node;
 }
 
-namespace OOModel {
-	class Method;
+namespace Visualization {
+	class Item;
+}
+
+namespace Export {
+	class SourceFile;
+	class TextToNodeMap;
 }
 
 namespace OODebug {
 
-class RunProcess;
+class CompilerMessage;
 
-class OODEBUG_API JavaRunner
+class OODEBUG_API JavaCompiler
 {
 	public:
-		/**
-		 * Finds a main method in the tree and runs the Programm from this main method.
-		 * If there is a valid main method the pointer to this method is returned.
-		 */
-		static OOModel::Method* runTree(Model::TreeManager* manager, const QString& pathToProjectContainerDirectory,
-								  bool debug = false);
+		static void compileTree(Model::TreeManager* manager, const QString& pathToProjectContainerDirectory,
+										bool includeDebugSymbols = false);
 
 	private:
-		static void noMainMethodWarning(Model::Node* node);
-		static void handleOutput();
-		static void handleErrorOutput();
-
-		static void addConsole(Model::Node* node);
-
-		static RunProcess& runProcess();
+		static void visualizeMessage(Visualization::Item* item, Model::Node* node,
+											  const QString& message, const QString& type = "default");
 };
 
 } /* namespace OODebug */
